Replaces CommandMocks with CommandMockUtils, UpdateMockUtils
