+++ bot/src/test/java/edu/java/bot/command/AuthorizedCommandTest.java
@@ -3,6 +3,7 @@ package edu.java.bot.command;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import edu.java.bot.entity.User;
+import edu.java.bot.mock.UpdateMockUtils;
 import edu.java.bot.repository.UserRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
@@ -22,7 +23,7 @@ class AuthorizedCommandTest {
         when(userRepository.findUserById(2L)).thenReturn(Optional.empty());
         when(authorizedCommand.getUserRepository()).thenReturn(userRepository);
 
-        Update updateMock = CommandMocks.getUpdateMock(1L, 2L);
+        Update updateMock = UpdateMockUtils.getUpdateMock(1L, 2L);
         SendMessage actualMessage = authorizedCommand.handle(updateMock);
 
         SendMessage expectedMessage = new SendMessage(1L, "User is not registered.");
@@ -38,7 +39,7 @@ class AuthorizedCommandTest {
         SendMessage expectedMessage = new SendMessage(1L, "Test message");
         when(authorizedCommand.authorizedHandle(any(Update.class), any(User.class))).thenReturn(expectedMessage);
 
-        Update updateMock = CommandMocks.getUpdateMock(1L, 2L);
+        Update updateMock = UpdateMockUtils.getUpdateMock(1L, 2L);
         SendMessage actualMessage = authorizedCommand.handle(updateMock);
 
         verify(authorizedCommand).authorizedHandle(any(Update.class), any(User.class));
