+++ bot/src/test/java/edu/java/bot/command/HelpCommandTest.java
@@ -3,6 +3,8 @@ package edu.java.bot.command;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.model.request.ParseMode;
 import com.pengrad.telegrambot.request.SendMessage;
+import edu.java.bot.mock.CommandMockUtils;
+import edu.java.bot.mock.UpdateMockUtils;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -30,11 +32,11 @@ class HelpCommandTest {
 
     @Test
     void handleWithCommands() {
-        Command command1 = CommandMocks.getCommandMock("/command", "description of command");
-        Command command2 = CommandMocks.getCommandMock("/another", "another description");
+        Command command1 = CommandMockUtils.getCommandMock("/command", "description of command");
+        Command command2 = CommandMockUtils.getCommandMock("/another", "another description");
         Command helpCommand = new HelpCommand(List.of(command1, command2));
 
-        Update updateMock = CommandMocks.getUpdateMock(1L);
+        Update updateMock = UpdateMockUtils.getUpdateMock(1L);
         SendMessage actualMessage = helpCommand.handle(updateMock);
 
         SendMessage expectedMessage = new SendMessage(1L, """
@@ -49,7 +51,7 @@ class HelpCommandTest {
     void handleWithoutCommands() {
         Command helpCommand = new HelpCommand(List.of());
 
-        Update updateMock = CommandMocks.getUpdateMock(1L);
+        Update updateMock = UpdateMockUtils.getUpdateMock(1L);
         SendMessage actualMessage = helpCommand.handle(updateMock);
 
         SendMessage expectedMessage = new SendMessage(1L, "")
@@ -63,7 +65,7 @@ class HelpCommandTest {
         Command helpCommand = new HelpCommand(commands);
         commands.add(helpCommand);
 
-        Update updateMock = CommandMocks.getUpdateMock(1L);
+        Update updateMock = UpdateMockUtils.getUpdateMock(1L);
         SendMessage actualMessage = helpCommand.handle(updateMock);
 
         SendMessage expectedMessage = new SendMessage(1L, """
