+++ bot/src/test/java/edu/java/bot/mock/CommandMockUtils.java
@@ -1,0 +1,18 @@
+package edu.java.bot.mock;
+
+import edu.java.bot.command.Command;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class CommandMockUtils {
+
+    public static Command getCommandMock(String commandName, String description) {
+        Command command = mock(Command.class);
+
+        // stubbing
+        when(command.command()).thenReturn(commandName);
+        when(command.description()).thenReturn(description);
+
+        return command;
+    }
+}
